@@ -8,16 +8,15 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
-    "react": "^16.13.1",
+    "framer-motion": "^10.12.16",
+    "react": "^18.2.0",
     "react-animated-cursor": "^2.5.2",
-    "react-dom": "^16.13.1",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
     "react-router-dom": "5.2.0",
     "react-scripts": "^5.0.1",
     "react-transition-group": "4.4.1",
-    "styled-components": "^5.1.1",
-    "swiper": "6.5.7",
-    "uuid": "8.3.2"
+    "styled-components": "^5.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
