@@ -1,12 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from "react-dom/client";
 import App from './App';
 import Typography from './styles/Typography'
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("root"))
+
+root.render(
     <>
         <Typography />
         <App />
-    </>,
-    document.getElementById('root')
-);
+    </>
+)
+
